crypto/Discrete RNG/deploy/requirements.txt
pycryptodome
fastcrc